+++ src/models/__tests__/User.spec.ts
@@ -1,5 +1,7 @@
 import 'jest-expect-message';
-import { DEFAULT_ORCID_URL, LogInType, User, UserRole } from '../User';
+import { generalConfig } from "../../config/generalConfig";
+import { normaliseHttpProtocol } from "../../utils/helpers";
+import { LogInType, User, UserRole } from '../User';
 import bcrypt from 'bcryptjs';
 import casual from 'casual';
 import { logger } from '../../__mocks__/logger';
@@ -25,7 +27,7 @@ describe('constructor', () => {
       role: UserRole.ADMIN,
       givenName: casual.first_name,
       surName: casual.last_name,
-      orcid: `${DEFAULT_ORCID_URL}0000-0000-0000-000X`,
+      orcid: normaliseHttpProtocol(`${generalConfig.orcidBaseURL}0000-0000-0000-000X`),
       ssoId: casual.uuid,
       languageId: lang.id,
     }
@@ -83,7 +85,7 @@ describe('prepForSave standardizes the format of properties', () => {
       givenName: ' Test ',
       surName: '  user  ',
       languageId: 'test',
-      orcid: `${DEFAULT_ORCID_URL}0000-0000-0000-000X`,
+      orcid: `${generalConfig.orcidBaseURL}0000-0000-0000-000X`,
     });
     user.prepForSave();
     expect(user.email).toEqual('TESTer@exaMPle.cOm');
@@ -91,7 +93,7 @@ describe('prepForSave standardizes the format of properties', () => {
     expect(user.surName).toEqual('User');
     expect(user.role).toEqual(UserRole.RESEARCHER);
     expect(user.languageId).toEqual(defaultLanguageId);
-    expect(user.orcid).toEqual(`${DEFAULT_ORCID_URL}0000-0000-0000-000X`);
+    expect(user.orcid).toEqual(normaliseHttpProtocol(`${generalConfig.orcidBaseURL}0000-0000-0000-000X`));
   });
 });
 
@@ -114,10 +116,10 @@ describe('prepForSave properly handles ORCIDs', () => {
       givenName: ' Test ',
       surName: '  user  ',
       languageId: 'test',
-      orcid: `${DEFAULT_ORCID_URL.replace(/https?:\/\//, '')}0000-0000-0000-000X`,
+      orcid: `${generalConfig.orcidBaseURL.replace(/https?:\/\//, '')}0000-0000-0000-000X`,
     });
     user.prepForSave();
-    expect(user.orcid).toEqual(`${DEFAULT_ORCID_URL}0000-0000-0000-000X`);
+    expect(user.orcid).toEqual(normaliseHttpProtocol(`${generalConfig.orcidBaseURL}0000-0000-0000-000X`));
   });
 
   it('should handle the ORCID ID without base URL', () => {
@@ -129,25 +131,7 @@ describe('prepForSave properly handles ORCIDs', () => {
       orcid: `0000-0000-0000-000X`,
     });
     user.prepForSave();
-    expect(user.orcid).toEqual(`${DEFAULT_ORCID_URL}0000-0000-0000-000X`);
-  });
-});
-
-describe('formatORCID', () => {
-  // Test the ORCID formatting
-  it('should return null for an invalid ORCID', () => {
-    expect(User.formatORCID('25t24g45g45g546gt')).toBeNull();
-    expect(User.formatORCID('0000-0000-0000')).toBeNull();
-    expect(User.formatORCID(`${DEFAULT_ORCID_URL}/0000-0000-000`)).toBeNull();
-  });
-
-  it('should return the ORCID with the default URL', () => {
-    expect(User.formatORCID('0000-0000-0000-000X')).toEqual(`${DEFAULT_ORCID_URL}0000-0000-0000-000X`);
-  });
-
-  it('should return the ORCID as is if it already a valid ORCID with the base URL', () => {
-    expect(User.formatORCID(`${DEFAULT_ORCID_URL}0000-0000-0000-000X`)).toEqual(`${DEFAULT_ORCID_URL}0000-0000-0000-000X`);
-    expect(User.formatORCID('https://sandbox.orcid.org/0000-0000-0000-000X')).toEqual('https://sandbox.orcid.org/0000-0000-0000-000X');
+    expect(user.orcid).toEqual(normaliseHttpProtocol(`${generalConfig.orcidBaseURL}0000-0000-0000-000X`));
   });
 });