+++ src/services/commonStandardService.ts
@@ -1,6 +1,6 @@
-import { DMPHubConfig } from "../config/dmpHubConfig";
-import { generalConfig } from "../config/generalConfig";
-import { MyContext } from "../context";
+import {DMPHubConfig} from "../config/dmpHubConfig";
+import {generalConfig} from "../config/generalConfig";
+import {MyContext} from "../context";
 import {
   DMPCommonStandard,
   DMPCommonStandardAffiliation,
@@ -24,12 +24,13 @@ import { defaultLanguageId } from "../models/Language";
 import { Plan, PlanVisibility } from "../models/Plan"
 import { Project } from "../models/Project";
 import { RelatedWork } from "../models/RelatedWork";
-import { ORCID_REGEX, User } from "../models/User";
+import { User } from "../models/User";
 import { VersionedTemplate } from "../models/VersionedTemplate";
-import { isNullOrUndefined, valueIsEmpty } from "../utils/helpers";
+import { isNullOrUndefined, valueIsEmpty, formatORCID, ORCID_REGEX } from "../utils/helpers";
 import { ResearchDomain } from "../models/ResearchDomain";
 import { Answer } from "../models/Answer";
 import { PlanContributor } from "../models/Contributor";
+import { ExternalContributor } from "../types";
 
 // eslint-disable-next-line no-useless-escape
 export const DOI_REGEX = /^(https?:\/\/)?(doi\.org\/)?(doi:)?(10\.\d{4,9}\/[-._;()/:\w]+)$/;
@@ -543,4 +544,40 @@ export const loadNarrativeTemplateInfo = async (
   narrative.sections.sort((a, b) => a.sectionOrder - b.sectionOrder);
 
   return narrative;
+}
+
+// Parses a contributor into a standard format
+export function parseContributor(
+  input: DMPCommonStandardContact | DMPCommonStandardContributor | null,
+): ExternalContributor | null {
+  if (isNullOrUndefined(input)) {
+    return null;
+  }
+
+  const nameParts = input.name.split(",").map((t) => t.trim());
+  const surName = nameParts[0];
+  const givenName = nameParts[nameParts.length - 1];
+  const email = input.mbox;
+  let orcid = null;
+
+  // ORCID ID
+  if ("contact_id" in input && input.contact_id.type == "orcid") {
+    orcid = formatORCID(input.contact_id.identifier);
+  } else if ("contributor_id" in input && input.contributor_id.type == "orcid") {
+    orcid = formatORCID(input.contributor_id.identifier);
+  }
+
+  // ROR ID
+  let rorId = null;
+  if (input.dmproadmap_affiliation?.affiliation_id?.type == "ror") {
+    rorId = input.dmproadmap_affiliation.affiliation_id.identifier;
+  }
+
+  return {
+    affiliationId: rorId,
+    givenName: givenName,
+    surName: surName,
+    orcid: orcid,
+    email: email,
+  };
 }