+++ src/utils/helpers.ts
@@ -4,6 +4,22 @@
 import { formatISO9075 } from "date-fns";
 import { generalConfig } from "../config/generalConfig";
 
+export const ORCID_REGEX = /^(https?:\/\/)?(www\.)?(sandbox\.)?(orcid\.org\/)?([0-9]{4}-[0-9]{4}-[0-9]{4}-[0-9]{3}[0-9X])$/;
+
+// Ensure that the ORCID is in the correct format (https://orcid.org/0000-0000-0000-0000)
+export function formatORCID(orcidIn: string): string {
+  // If it is blank or already in the correct format, return it
+  if (!valueIsEmpty(orcidIn) && (orcidIn.match(ORCID_REGEX) && orcidIn.startsWith('http'))) return normaliseHttpProtocol(orcidIn);
+
+  // If it matches the ORCID format but didn't start with http then its just the id
+  if (!valueIsEmpty(orcidIn) && orcidIn.match(ORCID_REGEX)) {
+    return normaliseHttpProtocol(`${generalConfig.orcidBaseURL}${orcidIn.split('/').pop()}`);
+  }
+
+  // Otherwise it's not an ORCID
+  return null;
+}
+
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 export function stringToArray(array: any, delimiter = ' ', defaultResponse: string[] = []): string[] {
   if (typeof array === 'string') {
@@ -35,14 +51,27 @@ export function capitalizeFirstLetter(str: string): string {
 }
 
 // Remove know Protocol and Domain portions of identifiers from the string
+// Handles removal of http:// and https:// protocols by removing them from string and base URLs
 export function stripIdentifierBaseURL(str: string): string {
   if (!str) return '';
 
-  return str.replace(generalConfig.dmpIdBaseURL, '')
-            .replace(generalConfig.orcidBaseURL, '')
-            .replace(generalConfig.rorBaseURL, '')
-            .replace(/^\//, '')
-            .trim();
+  const normalisedStr = normaliseHttpProtocol(str);
+  for (const baseUrl of [
+    generalConfig.dmpIdBaseURL,
+    generalConfig.orcidBaseURL,
+    generalConfig.rorBaseURL,
+  ]) {
+    const normalisedBase = normaliseHttpProtocol(baseUrl);
+    if (normalisedStr.startsWith(normalisedBase)) {
+      return normalisedStr
+        .slice(normalisedBase.length)
+        .trim();
+    }
+  }
+
+  return str
+    .replace(/^\//, '') // Strip leading /
+    .trim();
 }
 
 // Verify that a string is a valid identifier
@@ -119,3 +148,37 @@ export function getCurrentDate(): string {
 export function randomHex(size: number): string {
   return [...Array(size)].map(() => Math.floor(Math.random() * 16).toString(16)).join('');
 }
+
+// Normalises dates
+export function normaliseDate (date: string | null): string {
+  if (date === null || date === undefined) {
+    return null;
+  }
+
+  // Split into parts, convert to integers and only continue if 3 integer parts
+  const parts = date
+    .split("-")
+    .map((p) => parseInt(p.trim(), 10))
+    .filter((p) => Number.isInteger(p));
+  if (parts.length !== 3) {
+    return null;
+  }
+
+  // Convert to numbers
+  return `${parts[0]}-${String(parts[1]).padStart(2, "0")}-${String(parts[2]).padStart(2, "0")}`;
+}
+
+export function stripHttpProtocol(input: string | null) {
+  if (isNullOrUndefined(input)) {
+    return null;
+  }
+  return input.trim().replace(/^https?:\/\//, '');
+}
+
+export function normaliseHttpProtocol(input: string | null) {
+  if (isNullOrUndefined(input)) {
+    return null;
+  }
+  return input.trim().replace(/^http:\/\//, 'https://');
+}
+